fix_: move version to separate package
