+++ .gitignore
@@ -120,8 +120,8 @@ bindata.go
 migrations.go
 cmd/status-backend/server/endpoints.go
 protocol/messenger_handlers.go
-params/VERSION
-params/GIT_COMMIT
+internal/version/VERSION
+internal/version/GIT_COMMIT
 
 # Don't ignore generated files in the vendor/ directory
 !vendor/**/*.pb.go
