+++ internal/version/version.go
@@ -1,4 +1,4 @@
-package params
+package version
 
 import (
 	_ "embed"
@@ -8,7 +8,7 @@ import (
 // Use go:generate script to get the version and git commit.
 // VERSION and GIT_COMMIT files are used in further `go:embed` commands to load values to the variables.
 // Suppress errors, assuming files have already been properly generated. Required for Docker builds.
-//go:generate sh -c "../_assets/scripts/version.sh > VERSION || true"
+//go:generate sh -c "../../_assets/scripts/version.sh > VERSION || true"
 //go:generate sh -c "git rev-parse --short HEAD > GIT_COMMIT || true"
 
 var (
@@ -21,9 +21,6 @@ var (
 	//go:embed GIT_COMMIT
 	gitCommit string
 )
-
-// IpfsGatewayURL is the Gateway URL to use for IPFS
-const IpfsGatewayURL = "https://ipfs.status.im/"
 
 func init() {
 	version = strings.TrimSpace(version)
